fix: append slash to the end of ingress path
There is a js file named `common.es-2015.xxx.js` in the portal of the
harbor 2.3.0, the request of this file will be transported to the core
component when there isn't a slash at the end of the ingress path,
the portal will not be accessible. So we need to append a slash to the end
of the ingress path.

Signed-off-by: He Weiwei <hweiwei@vmware.com>